pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "legal-kg"
version = "1.0.0"
description = "Legal Knowledge Graph -- AI-powered graph querying for legal documents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Graph
    "neo4j>=5.18",
    # API
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    # Config & validation
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    # LLM providers (core — OpenAI ships by default)
    "openai>=1.30",
    # Utilities
    "python-dotenv>=1.0",
    "httpx>=0.27",
    # PDF ingestion
    "pdfplumber>=0.11",

]

[project.optional-dependencies]
# Install the provider(s) you want to use:
#   pip install -e ".[anthropic]"
#   pip install -e ".[gemini]"
#   pip install -e ".[anthropic,gemini]"
anthropic = ["anthropic>=0.28"]
gemini    = ["google-generativeai>=0.8"]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
]
all-providers = [
    "anthropic>=0.28",
    "google-generativeai>=0.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
